move QuantKey to quant_utils
