+++ vllm/v1/attention/backends/flashinfer.py
@@ -17,9 +17,10 @@ from flashinfer.utils import FP4Tensor
 from vllm import _custom_ops as ops
 from vllm.attention.backends.abstract import (AttentionBackend, AttentionImpl,
                                               AttentionType)
-from vllm.compilation.fusion import QuantKey, kFp8StaticTensorSym, kNvfp4Quant
+from vllm.compilation.fusion import kFp8StaticTensorSym, kNvfp4Quant
 from vllm.config import CUDAGraphMode, VllmConfig
 from vllm.logger import init_logger
+from vllm.model_executor.layers.quantization.utils.quant_utils import QuantKey
 from vllm.platforms import current_platform
 from vllm.utils import cdiv, is_pin_memory_available
 from vllm.utils.flashinfer import (supports_trtllm_attention,
